[feat] LH- 04.06.21-feature/PROT-1009 	Move trade tasks to stats bot
